major merge attempt of 46/47/48/50
